+++ arroyo/processing/strategies/__init__.py
@@ -3,12 +3,16 @@ from arroyo.processing.strategies.abstract import (
     ProcessingStrategy,
     ProcessingStrategyFactory,
 )
-from arroyo.processing.strategies.batching import BatchStep, UnbatchStep, ValuesBatch
+from arroyo.processing.strategies.batching import BatchStep, UnbatchStep
 from arroyo.processing.strategies.collect import CollectStep, ParallelCollectStep
 from arroyo.processing.strategies.commit import CommitOffsets
 from arroyo.processing.strategies.filter import FilterStep
 from arroyo.processing.strategies.produce import Produce
-from arroyo.processing.strategies.run_task import RunTaskInThreads
+from arroyo.processing.strategies.run_task import (
+    RunTask,
+    RunTaskInThreads,
+    RunTaskWithMultiprocessing,
+)
 from arroyo.processing.strategies.transform import ParallelTransformStep, TransformStep
 
 __all__ = [
@@ -22,8 +26,9 @@ __all__ = [
     "ProcessingStrategy",
     "ProcessingStrategyFactory",
     "Produce",
+    "RunTask",
     "RunTaskInThreads",
     "BatchStep",
     "UnbatchStep",
-    "ValuesBatch",
+    "RunTaskWithMultiprocessing",
 ]